Safari-609.1.11.1/LayoutTests/fast/selectors/map-tag-default-display-expected.txt

Verify default display of <map>.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS getComputedStyle(document.getElementById('target')).display is "inline"
PASS successfullyParsed is true

TEST COMPLETE
